Add spaced repetition settings

+++ composeApp/src/commonMain/kotlin/com/gaoyun/yanyou_kototomo/data/local/card/CardProgress.kt
@@ -12,7 +12,7 @@ data class CardProgress(
     val cardId: String,
     val lastReviewed: LocalDate,
     val interval: Int,
-    val easeFactor: Double,
+    val easeFactor: Float,
     val nextReview: LocalDate,
 )
